// Controls/_scroll/Container/Type.ts
/**
 * @kaizen_zone 7b560386-8131-481a-b9c0-8b3ede6f29a0
 */
export enum POSITION {
    TOP = 'top',
    BOTTOM = 'bottom',
    LEFT = 'left',
    RIGHT = 'right',
}

export enum SCROLL_MODE {
    VERTICAL = 'vertical',
    HORIZONTAL = 'horizontal',
    VERTICAL_HORIZONTAL = 'verticalHorizontal',
    NONE = 'none',
}

export enum ScrollShadowVisibility {
    Hidden = 'hidden',
    Visible = 'visible',
    Auto = 'auto',
    Gridauto = 'gridauto',
}

export enum OverscrollBehavior {
    Auto = 'auto',
    Contain = 'contain',
}